refactor: move Function_ to be Function
Also move old `Function` struct to `completion`
